requirements.txt
websocket_client==0.56.0
confluent-kafka==0.11.6
cassandra-driver==3.14.0
hdfs==2.5.8